requirements_test.txt
pytest
pytest-homeassistant-custom-component
ruff
scikit-learn
pandas
